GIRAPH-985: Add more metrics
Summary: When limit number of open requests is on, significant part of computation can be spent in just waiting. So adding metrics for total amount of time compute threads spent in waiting on open requests, and histogram with compute times per partition. These should help detecting why some job is slower than expected (is it from messaging or compute). Also changed the way we wait on open requests, so we continue on different limit than we stop, so we wouldn't have huge number of very tiny pauses.

Test Plan: Run some jobs on the cluster and looked through these metrics.

Reviewers: sergey.edunov, pavanka

Subscribers: ikabiljo

Differential Revision: https://reviews.facebook.net/D31683