[GIRAPH-1019] Optimizing and debugging vertex mutation mechanism
Summary:
The old implementation of vertex mutation mechanism was single-threaded and had some redundant computation. The single threaded behavior is causing a huge performance degradation for out-of-core case, since all the partitions are being read and written sequentially in one thread to apply mutations. Also, in case where a vertex is mutated and has messages at the same time, the current code fails to execute which does not seem to be the expected behavior.

This diff implements an optimized multi-threaded approach for vertex mutations. With this diff, vertex mutation happens at the beginning of processing each partition. Also, parts of partition migration code is modified to migrate mutations as well.

Test Plan: mvn clean verify

Reviewers: avery.ching, maja.kabiljo, dionysis.logothetis

Reviewed By: dionysis.logothetis

Differential Revision: https://reviews.facebook.net/D40821