[GIRAPH-1022] Adaptive out-of-core mechanism for input superstep and graph
Summary: This code adds the ability to adaptively control the out-of-core mechanism for graph data structure at run-time during input/output superstep and computation superstep. Basically, the implemented mechanism monitors the amount of available free memory in a separate thread. If there is not enough memory, the code adjusts the number of partitions in memory, and spills a series of partitions/buffers to disk. Also, if the amount of free memory is more than expected, some of the on-disk partitions are brought back to memory. Additionally, if amount of free memory is marginal, the mechanism mocks the memory usage by gradually bringing partitions to memory.

Test Plan:
mvn clean verify
Unit tests added to giraph-core
End-to-end test added to giraph-example
Running the code on PageRank on a large graph and not getting OOM failures.

Reviewers: maja.kabiljo, sergey.edunov, avery.ching, dionysis.logothetis

Reviewed By: dionysis.logothetis

Differential Revision: https://reviews.facebook.net/D40563