[GIRAPH-1023] Adding out-of-core messages to previously implemented adaptive out-of-core mechanism
Summary:
This is the continuation of the previous diff on out-of-core mechanism. This diff completes the last diff by adding out-of-core messages, making the entire out-of-core mechanism a cohesive entity in Giraph.

This diff also improves the API of PartitionStore by some minor refactoring.

Test Plan:
mvn clean verify
running pagerank and turning message combiner off on a large graph with limited memory does not fail

Reviewers: maja.kabiljo, sergey.edunov, avery.ching, dionysis.logothetis

Reviewed By: dionysis.logothetis

Differential Revision: https://reviews.facebook.net/D42897