Merge seeded and unseeded BFS into a single BFS implementation
Summary: This change creates a general Breadth First Search version which supports the default BFS where distances to one or more seeds are computed. Additionally, this new version also supports assigning vertices to closest seeds for the purpose of clustering the vertices. This change provides a BlockFactory which highlights this functionality in addition to test cases.

Test Plan: Test cases for the new functionality added.

Reviewers: spupyrev, ikabiljo

Reviewed By: ikabiljo

Subscribers: dionysis.logothetis, sergey.edunov, maja.kabiljo

Differential Revision: https://reviews.facebook.net/D47985