GIRAPH-1035: Make sure we are able to use all compute threads
Summary: The default logic of choosing the number of partitions when we use few workers and a lot of compute threads ends up choosing less partitions than there are threads. Add additional setting to prevent that.

Test Plan: Run a job with a few workers and lot of threads and verified number of partitions is set properly. mvn verify passed.

Differential Revision: https://reviews.facebook.net/D48993