GIRAPH-1052: Fix makeSymmetricUnweighted
Summary: PrepareGraphPieces.makeSymmetricUnweighted is currently very inefficient for skewed degree graphs, because it reuses set objects based on the number of in edges, but also adds all out edges to the set, so sets which should be small can become huge. Since incoming ids are unique anyways, we don't need to add them to the set.

Test Plan: Ran a job without and with the change, verified result is the same but it's much faster now

Reviewers: ikabiljo

Differential Revision: https://reviews.facebook.net/D56481