[GIRAPH-1053] Log exceptions to command line
Summary: When we know an exception occurred, log it to command line to make it easier for people running jobs to see what the issue was.

Test Plan: Ran two jobs, one with error in input one with error in compute, verified exception is printed to command line. Also ran a normal job and verified it didn't print anything new to command line

Differential Revision: https://reviews.facebook.net/D56931