GIRAPH-1058: Fix connection retry logic
Summary: Currently when we fail to connect to a channel we retry immediately and that retry most often fails. Add a short wait between retries, and improve the check for whether the channel connected successfully.

Test Plan: Ran multiple jobs which were often failing before the fix, with fix they worked

Differential Revision: https://reviews.facebook.net/D57447