GIRAPH-1068 Make Zookeeper accept 0 as a port number and let it choose any available free port
Summary:
We have a few use cases where having zookeeper bound to specific port is very inconvenient.
1) Unit tests that run in parallel.
2) Shared clusters where multiple giraph instances can run on the same machines.

In theory we don't need to know what port zookeeper will run on. In most cases we're fine with any port available.
Picking any available port is currently supported by the server socket, but is not supported in the code that parses zookeper configs (this code lives in zookeper).
We don't have to parse configs though, as we have a way to run zookeper in process. And in that case we can have a full control on how zookeeper is initialized.

For this task I want to allow 0 as a port number for zookeeper. Which will allow us to run zookeeper on any available port. And I will also remove "out of process" zookeeper, as it clearly provides no benefits to us.

Note: it will still be possible to run external zookeper, if you have it running somewhere as a service.

Note2: this change is intended to remove the functionality to support multiple zk servers running as a part of the Giraph job and only support a single zk server. If you want to run multiple zookeepers, you need to configure them separately and let Giraph use existing zookeper quorum

Test Plan:
tested a few things:
picking any available port
-Dgiraph.zkServerPort=0
using external zookeeper:
-Dgiraph.zkList="hadoopXXX.YYY.facebook.com:22181"
using specified port:
-Dgiraph.zkServerPort=22128

Reviewers: majakabiljo, maja.kabiljo, dionysis.logothetis, avery.ching, heslami

Reviewed By: avery.ching, heslami

Differential Revision: https://reviews.facebook.net/D59109