GIRAPH-1083: Make sure we fail after exception in ooc-io thread happens
Summary: Currently if some exception happens in ooc-io thread the job is left running for long time after the exception. We should make sure we fail early.

Test Plan: Ran a job with ooc on where I simulated the failure, without change job hangs for a long time, with the change it fails right after the exception happens, and logs it to command line.

Differential Revision: https://reviews.facebook.net/D60291