GIRAPH-1087: Retry requests after channel failure
Summary: We currently don't have a callback to retry requests after channel failure, and would either wait for request timeout or not retrying request at all at places where we don't wait for open requests.

Test Plan: Hard to reproduce the issue (ran many jobs but was unable to), we'll see if the problem happens again in prod with this change.

Differential Revision: https://reviews.facebook.net/D60675