GIRAPH-1097 Fix TestOutOfCore.testOutOfCoreLocalDiskAccessor
Summary:
On my laptop it failed because of an NPE in WorkerSuperstepMetrics.
I tracked it down and found that it is triggered from the branch of code that prints out metrics. We don't normally print out metrics in unit tests, so I'd expect this feature doesn't exist or not functional in hadoop_1. I'll try to disable it, to see how jenkins reacts.

Test Plan:  mvn test -pl giraph-examples -am -Dtest=TestOutOfCore -DfailIfNoTests=false -Phadoop_1

Reviewers: maja.kabiljo, dionysis.logothetis, heslami

Reviewed By: heslami

Differential Revision: https://reviews.facebook.net/D60873