GIRAPH-1104: NegativeArraySize exception in BigDataOutput
Summary:
BigDataIO is not properly handling large byte[] being written to it. Chunk them up when needed to respect the max single data output size.
With D61791 job was still failing with the same exception.

Test Plan: The job which was failing because of large edges now works, added a test

Differential Revision: https://reviews.facebook.net/D61839