GIRAPH-1211: Make retrying to send network requests after timeout optional
Summary: Using counters added in GIRAPH-1205 we were able to confirm that resending network requests after timeout almost never succeeds, so add an option to fail early instead of keep trying to resend these network requests indefinitely.

Test Plan: Made response not being sent for a specific request id, verified that with keeping default value for this new option we kept trying to resend the request, and when making it false the job failed when request timeout was noticed.