GIRAPH-1251: Add optional SSLHandler for all Netty Communication
With this change its possible to encrypt all communication via Netty
with SSL Handler. It also exposes function to add custom authn and
authz checks on tls handshake complete. Once this flag is set, all
Netty traffic will be ensured to be encrypted.

Testing:

- mvn clean verify passes.
- When the sslEncrypt flag is off there were no changes to the existing workflow.
- On enabling the flag (sslEncrypt) and having a custom
SSL Event Handler, there was no performance hit. A standard set of tests were
run with and without this flag in the same enviornment and the results showed
-5% to 5% difference in total CPU time utilized by the jobs. This is within
acceptable limits and is normal variance.